Proposed (non-working) change to pull CrudOperations out.
