+++ src/CMakeLists.txt
@@ -17,6 +17,7 @@ add_subdirectory(third_party)
 # keep in alpha order
 add_subdirectory(canaries)
 add_subdirectory(cast_core)
+add_subdirectory(crud_operations)
 add_subdirectory(driver)
 add_subdirectory(gennylib)
 add_subdirectory(metrics)
@@ -26,7 +27,7 @@ add_subdirectory(value_generators)
 add_subdirectory(workloads)
 
 # This makes the project importable from the build directory
-export(TARGETS value_generators metrics gennylib loki cast_core poplarlib
+export(TARGETS value_generators crud_operations metrics gennylib loki cast_core poplarlib
        FILE GennyLibraryConfig.cmake)
 
 # Below generates a cmake module so if you make install from this
